fix(exec): force owner-only mode when rewriting an existing tfvars file
os.WriteFile applies perm only on create, so a leftover 0644 from an
older terragraph stayed world-readable for the whole run. WriteTFVars
now removes the path first, and the permission assertions live under
//go:build !windows where mode bits are meaningful.
